Resolucion Ej6 en Assembler

+++ adm_c16/Core/Src/Ejercicio6.c
@@ -12,7 +12,7 @@ void pack32to16 (int32_t * vectorIn, int16_t *vectorOut, uint32_t longitud)
 {
 	int i = 0;
 	for (i = 0; i < longitud; i++) {
-		*(vectorOut + i) = ((*(vectorIn + i)) >> 16) & 0xff;
+		*(vectorOut + i) = (uint16_t) (((*(vectorIn + i)) >> 16) & 0xff);
 	}
 }
